Main capabilities back. added support to open other tree and traverse them. no checks yet
